+++ examples/next-js/src/app/api/actions/transfer-sol/route.ts
@@ -8,6 +8,7 @@ import {
   ActionGetResponse,
   ActionPostRequest,
   createActionHeaders,
+  ActionError,
 } from "@solana/actions";
 import {
   clusterApiUrl,
@@ -33,6 +34,7 @@ export const GET = async (req: Request) => {
     ).toString();
 
     const payload: ActionGetResponse = {
+      type: "action",
       title: "Actions Example - Transfer Native SOL",
       icon: new URL("/solana_devs.jpg", requestUrl.origin).toString(),
       description: "Transfer SOL to another Solana wallet",
@@ -71,9 +73,9 @@ export const GET = async (req: Request) => {
     });
   } catch (err) {
     console.log(err);
-    let message = "An unknown error occurred";
-    if (typeof err == "string") message = err;
-    return new Response(message, {
+    let actionError: ActionError = { message: "An unknown error occurred" };
+    if (typeof err == "string") actionError.message = err;
+    return Response.json(actionError, {
       status: 400,
       headers,
     });
@@ -82,9 +84,7 @@ export const GET = async (req: Request) => {
 
 // DO NOT FORGET TO INCLUDE THE `OPTIONS` HTTP METHOD
 // THIS WILL ENSURE CORS WORKS FOR BLINKS
-export const OPTIONS = async (req: Request) => {
-  return new Response(null, { headers });
-};
+export const OPTIONS = async () => Response.json(null, { headers });
 
 export const POST = async (req: Request) => {
   try {
@@ -98,10 +98,7 @@ export const POST = async (req: Request) => {
     try {
       account = new PublicKey(body.account);
     } catch (err) {
-      return new Response('Invalid "account" provided', {
-        status: 400,
-        headers,
-      });
+      throw 'Invalid "account" provided';
     }
 
     const connection = new Connection(
@@ -158,9 +155,9 @@ export const POST = async (req: Request) => {
     });
   } catch (err) {
     console.log(err);
-    let message = "An unknown error occurred";
-    if (typeof err == "string") message = err;
-    return new Response(message, {
+    let actionError: ActionError = { message: "An unknown error occurred" };
+    if (typeof err == "string") actionError.message = err;
+    return Response.json(actionError, {
       status: 400,
       headers,
     });